requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Werkzeug==3.1.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-Bootstrap==3.3.7.1
Flask-CKEditor==1.0.0
Flask-WTF==1.2.2
WTForms==3.2.1
python-dotenv==1.0.1